Resolved to do's
Resolved to do's by moving hard coded stuff to settings and for it to be allowed variable is input arguments. This introduced new classes, which got to do's also. So all in all I ended up with more to do's for the report builder and almost all of them being writing unit tests and refactor. However, a quick manual test provided the desired results for the reports. So the settings and the passing in of arguments work.